backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.0.13
django-auth-adfs==1.15.0
django-cors-headers==4.4.0
django-filter==24.3
django-mssql-backend==2.8.1
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
git-filter-repo==2.47.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mssql-django==1.5
pycparser==2.22
PyJWT==2.9.0
pyodbc==5.2.0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
